Desafio 4/requirements.txt
atomicwrites==1.1.5
attrs==18.1.0
certifi==2018.4.16
more-itertools==4.2.0
pluggy==0.6.0
py==1.5.3
pytest==3.6.2
six==1.11.0
PyJWT==1.7.1
